package/infra/pyproject.toml
[project]
name = "diopside-infra"
version = "0.1.0"
description = "AWS CDK infrastructure for Diopside"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib==2.201.0",
    "boto3>=1.38.36",
    "constructs==10.4.2",
]

[dependency-groups]
infra = [
    "aws-cdk-lib==2.201.0",
    "boto3>=1.38.36",
    "constructs==10.4.2",
]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.11.13",
    "mypy>=1.16.0",
]
layer = [
    "aws-lambda-powertools>=3.14.0",
    "fastapi>=0.115.12",
    "mangum>=0.19.0",
    "pydantic>=2.11.7",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
ignore = [
    "F401", # unused-import
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
exclude = ["../cdk.out"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
